crash on button multi-touch fixed + start commenting
added a line to CameraFragment and CameraConfirmFragment to prevent the
user from crashing the app if the send or shoot buttons are hit more
then once. Also, started adding comments to the top of files so anybody
in the future will be able to tell what is implemented in that file.